+++ toppy/client.py
@@ -3,12 +3,12 @@ from __future__ import annotations
 import asyncio
 from abc import abstractmethod
 from functools import wraps
-from typing import TYPE_CHECKING, AsyncGenerator, Optional, Type, Union
+from typing import TYPE_CHECKING, AsyncGenerator, ClassVar, Optional, Type
 
 import aiohttp
 
 from .errors import ClientNotReady, HTTPException
-from .http import BaseHTTPClient, DBLHTTPClient, TopGGHTTPClient
+from .http import BaseHTTPClient, DiscordBotListHTTPClient, DiscordBotsGGHTTPClient, TopGGHTTPClient
 from .utils import MISSING
 
 if TYPE_CHECKING:
@@ -18,7 +18,8 @@ if TYPE_CHECKING:
 
 __all__ = (
     'Client',
-    'DBLClient',
+    'DiscordBotListClient',
+    'DiscordBotsGGClient',
     'TopGGClient'
 )
 
@@ -99,9 +100,8 @@ class BaseClient:
         self.task.cancel()
 
     async def _post_stats_handler(self, bot_id: int, **kwargs) -> None:
-        resp = await self.http.post_stats(bot_id, **kwargs)
         try:
-            resp.raise_for_status()
+            await self.http.post_stats(bot_id, **kwargs)
         except HTTPException as exc:
             self.client.dispatch(f'{self.shortened}_post_error', exc)
         else:
@@ -118,7 +118,7 @@ class BaseClient:
         raise NotImplementedError
 
 
-class DBLClient(BaseClient):
+class DiscordBotListClient(BaseClient):
     """A Client to access the Discord Bot List API. This includes auto-posting Discord Bot Stats.
 
     https://discordbotlist.com/
@@ -148,18 +148,18 @@ class DBLClient(BaseClient):
         Moved to toppy.client
     """
 
-    http_class = DBLHTTPClient
+    http_class = DiscordBotListHTTPClient
     shortened = 'dbl'
 
     if TYPE_CHECKING:
-        http: DBLHTTPClient
+        http: DiscordBotListHTTPClient
 
     async def post_stats(self) -> None:
         """Post your bots stats to Discord Bot List.
         All stats are automatically found and posted.
 
-        dispatches `dbl_autopost_error` with the argument :class:`toppy.HTTPException` or derived classes
-        or `dbl_autopost_success` with no arguments.
+        dispatches `dbl_post_error` with the argument :class:`toppy.HTTPException` or derived classes
+        or `dbl_post_success` with no arguments.
         """
 
         bot_id = self._get_bot_id()
@@ -167,6 +167,54 @@ class DBLClient(BaseClient):
         kwargs = {
             'voice_connections': len(self.client.voice_clients),
             'users': len(self.client.users or []),
+            'guilds': len(self.client.guilds or [])
+        }
+
+        await self._post_stats_handler(bot_id, **kwargs)
+
+
+class DiscordBotsGGClient(BaseClient):
+    """A Client to access the DiscordBotsGG. This includes auto-posting Discord Bot Stats.
+
+    https://discord.bots.gg/
+
+    Parameters
+    ----------
+    client: :class:`Client`
+        The Discord Bot instance. Any Client derived from :class:`discord.Client` or any other fork's `Client`.
+    token: :class:`str`
+        The token for DiscordBotsGG.
+    interval: Optional[:class:`float`]
+        The interval in seconds to auto-post the stats.
+        Defaults to 600.
+    start_on_ready: :class:`bool`
+        Whether to start the auto post task when the bot is ready.
+        If False then it must be manually started with `start`.
+        Defaults to True.
+    session: Optional[:class:`aiohttp.ClientSession`]
+        The session for the HTTP Client.
+
+
+    .. versionadded:: 1.6
+    """
+
+    http_class = DiscordBotsGGHTTPClient
+    shortened = 'dbgg'
+
+    if TYPE_CHECKING:
+        http: DiscordBotsGGHTTPClient
+
+    async def post_stats(self) -> None:
+        """Post your bots stats to DiscordBotsGG
+        All stats are automatically found and posted.
+
+        dispatches `dbgg_post_error` with the argument :class:`toppy.HTTPException` or derived classes
+        or `dbgg_post_success` with no arguments.
+        """
+
+        bot_id = self._get_bot_id()
+
+        kwargs = {
             'guilds': len(self.client.guilds or [])
         }
 
@@ -298,8 +346,8 @@ class TopGGClient(BaseClient):
         """Post your bots stats to Top.gg.
         All stats are automatically found and posted.
 
-        dispatches `topgg_autopost_error` with the argument :class:`toppy.HTTPException` or derived classes
-        or `topgg_autopost_success` with no arguments.
+        dispatches `topgg_post_error` with the argument :class:`toppy.HTTPException` or derived classes
+        or `topgg_post_success` with no arguments.
         """
 
         bot_id = self._get_bot_id()
@@ -316,7 +364,7 @@ class TopGGClient(BaseClient):
 
 class Client:
     """A class designed to handle both Top.gg and Discord Bot List.
-    For individual handling use :class:`DBLClient` or :class:`TopGGClient`
+    For individual handling use :class:`DiscordBotListClient` or :class:`TopGGClient`
 
     Parameters
     ----------
@@ -325,6 +373,8 @@ class Client:
         It must fit the :class:`ClientProtocol`.
     dbl_token: :class:`str`
         The authorization token for Discord Bot List.
+    dbgg_token: :class:`str`
+        The authentication token for DiscordBotsGG
     topgg_token: :class:`str`
         The authorization token for Top.gg.
     interval: Optional[:class:`float`]
@@ -342,7 +392,16 @@ class Client:
 
     .. versionchanged:: 1.5
         ``client`` is no longer positional only.
+
+    .. versionchanged:: 1.6
+        Add support for DiscordBotsGG
     """
+    clients: ClassVar[tuple] = (
+        ('dbl', DiscordBotListClient),
+        ('dbgg', DiscordBotsGGClient),
+        ('topgg', TopGGClient)
+    )
+
     def __init__(
             self,
             client: ClientProtocol,
@@ -351,205 +410,111 @@ class Client:
         self.client = client
 
         self._original_options = options
+        self.__session: aiohttp.ClientSession = MISSING
 
-        self.__dbl: DBLClient = MISSING
-        self.__topgg: TopGGClient = MISSING
+        self.__dbl: Optional[DiscordBotListClient] = None
+        self.__dbgg: Optional[DiscordBotsGGClient] = None
+        self.__topgg: Optional[TopGGClient] = None
 
-    @property
-    def intervals(self) -> tuple[float, float]:
-        """Returns the intervals of the autopost task. First the dbl, then the topgg.
+    def _init(self):
+        interval: int = self._original_options.get('interval', 600)
+        start_on_ready: bool = self._original_options.get('start_on_ready', True)
 
-        Returns
-        --------
-        tuple[:class:`float`]
-        """
-        return self.__dbl.interval, self.__topgg.interval
+        for name, cls in self.clients:
+            token = self._original_options.get(f'{name}_token')
 
-    @intervals.setter
-    def intervals(self, new: Union[float, tuple[Optional[float], Optional[float]]]) -> None:
-        """Set new intervals.
-         If you pass in :class:`float` both the dbl and topgg client's interval will be set to it.
-         If you pass in tuple[Optional[:class:`float`, Optional[:class:`float`]] then the first
-         will set the dbl client and the second will set the topgg client.
+            if not token:
+                continue
 
-         Raises
-         -------
-         :exc:`TypeError` If it didn't receive :class:`float`, :class:`int`, or :class:`tuple`
-         """
-        if isinstance(new, (float, int)):
-            self.__dbl.interval = new
-            self.__topgg.interval = new
-        elif isinstance(new, tuple):
-            dbl, topgg = new
+            kwargs = {
+                'interval': interval,
+                'start_on_ready': start_on_ready,
+                'session': self.__session
+            }
 
-            if dbl is not None:
-                self.__dbl.interval = dbl
-            if topgg is not None:
-                self.__topgg.interval = topgg
+            if 'post_shard_count' in cls.__init__.__annotations__:
+                kwargs['post_shard_count'] = kwargs.get('post_shard_count', False)
 
-        else:
-            raise TypeError(f'Expected `float` or `tuple[float, float]` not {new.__class__.__name__!r}')
+            client = cls(self.client, token, **kwargs)
+            setattr(self, f'__{name}', client)
 
-    def _merge_starts(self) -> None:
+    def _merge(self) -> None:
         old_start = self.client.start
         old_close = self.client.close
 
         # used over setup_hook for fork support
         @wraps(old_start)
         async def start(*args, **kwargs) -> None:
-            task = self.client.loop.create_task(old_start(*args, **kwargs))
-
-            session = self._original_options.get('session', aiohttp.ClientSession(loop=self.client.loop))
-
-            self.__dbl = DBLClient(
-                self.client, token=self._original_options['dbl_token'],
-                interval=self._original_options.get('interval'),
-                start_on_ready=self._original_options.get('start_on_ready', True),
-                session=session
-            )
-
-            self.__topgg = TopGGClient(
-                self.client, token=self._original_options['topgg_token'],
-                interval=self._original_options.get('interval'),
-                post_shard_count=self._original_options.get('post_shard_count', False),
-                start_on_ready=self._original_options.get('start_on_ready', True),
-                session=session
-            )
-
-            if self._original_options.get('start_on_ready', True):
-                self.__dbl.start()
-                self.__topgg.start()
-
-            await task
+            self.__session = aiohttp.ClientSession(loop=self.client.loop)
+            await old_start(*args, **kwargs)
 
         @wraps(old_close)
         async def close() -> None:
-            if not self.__dbl.http.session.closed:
-                await self.__dbl.http.session.close()
-            if not self.__topgg.http.session.closed:
-                await self.__topgg.http.session.close()
-
+            await self.__session.close()
             await old_close()
 
         self.client.start = start  # type: ignore # "Cannot assign to method"
         self.client.close = close  # type: ignore
 
+    def _get_clients(self) -> list[BaseClient]:
+        clients = []
+
+        for client, _ in self.clients:
+            instance = getattr(self, client)
+            if instance:
+                clients.append(instance)
+
+        return clients
+
     def start(self):
         """Starts the autopost task."""
-        self.__dbl.start()
-        self.__topgg.start()
+        for client in self._get_clients():
+            client.start()
 
     def cancel(self):
         """Cancels the task of auto posting stats."""
-        self.__dbl.cancel()
-        self.__topgg.cancel()
+        for client in self._get_clients():
+            client.cancel()
 
     @property
-    def dbl_task(self) -> asyncio.Task:
+    def dbl(self) -> Optional[DiscordBotListClient]:
         """
-        The :class:`asyncio.Task` object for dbl autopost.
-        """
-        return self.__dbl.task
-
-    @property
-    def topgg_task(self) -> asyncio.Task:
-        """
-        The :class:`asyncio.Task` object for topgg autopost.
-        """
-        return self.__topgg.task
-
-    @property
-    def dbl(self) -> DBLClient:
-        """
-        Access the interval Client for Discord Bot List.
+        Access the Client for Discord Bot List.
         
         Returns
         --------
-        :class:`DBLCLient`
+        Optional[:class:`DBLCLient`]
         """
         return self.__dbl
 
     @property
-    def topgg(self) -> TopGGClient:
+    def dbgg(self) -> Optional[DiscordBotsGGClient]:
         """
-        Access the interval Client for Top.gg.
+        Access the Client for DiscordBotsGG.
+
+        Returns
+        --------
+        Optional[:class:`DiscordBotsGGClient`]
+        """
+        return self.__dbgg
+
+    @property
+    def topgg(self) -> Optional[TopGGClient]:
+        """
+        Access the Client for Top.gg.
         
         Returns
         --------
-        :class:`TopGGClient`
+        Optional[:class:`TopGGClient`]
         """
         return self.__topgg
 
     async def post_stats(self) -> None:
-        """Post your bots stats to Discord Bot List and Top.gg.
+        """Post your bots stats to all websites with a token found.
         All stats are automatically found and posted."""
-        await asyncio.gather(self.__dbl.post_stats(), self.__topgg.post_stats())
+        tasks = [
+            self.client.loop.create_task(client.post_stats())
+            for client in self._get_clients()
+        ]
 
-    async def search_bots(self, query: str, *, limit: Optional[int] = None, offset: Optional[int] = None
-                          ) -> list[Bot]:
-        """Search up bots on Top.gg.
-
-        Parameters
-        -----------
-        query: :class:`str`
-            The query to use when searching.
-        limit: Optional[:class:`int`]
-            The limit of :class:`Bot`'s to return.
-            Keyword only.
-        offset: Optional[:class:`int`]
-            The amount of bots to skip in the results.
-            Keyword only.
-
-        Returns
-        --------
-        list[:class:`Bot`]
-        """
-        return await self.__topgg.search_bots(query=query, limit=limit, offset=offset)
-
-    async def search_one_bot(self, bot_id: int, /) -> Bot:
-        """Search a single bot on Top.gg.
-
-        Parameters
-        ----------
-        bot_id: :class:`int`
-            The ID to search.
-            Positional only.
-
-        Returns
-        --------
-        :class:`Bot`:
-        """
-        return await self.search_one_bot(bot_id)
-
-    async def last_1000_votes(self, bot_id: int = None, /) -> AsyncGenerator[User, None]:
-        """Get the last 1000 votes of a bot on Top.gg
-
-        Parameters
-        -----------
-        bot_id: Optional[:class:`int`]
-            The ID of the bot.
-            Defaults to the Bot initialized with.
-            Positional only.
-
-        Yields
-        -------
-        :class:`User`
-        """
-        return self.__topgg.last_1000_votes(bot_id)
-
-    async def check_if_voted(self, bot_id: Optional[int], user_id: int) -> bool:
-        """Check if a user has voted on a bot
-
-        Parameters
-        ----------
-        bot_id: Optional[:class:`int`]
-            The ID of the bot.
-            Defaults to the Bot initialized with.
-        user_id: :class:`int`
-            The ID of the user.
-
-        Returns
-        --------
-        :class:`bool`
-        """
-        return await self.__topgg.check_if_voted(bot_id, user_id)
+        await asyncio.gather(*tasks)